requirements.txt
alembic==1.5.7
certifi==2020.12.5
chardet==3.0.4
distlib==0.3.1
dython==0.6.5.post1
filelock==3.0.12
isodate==0.6.0
itsdangerous==2.0.0
jedi==0.17.0
Jinja2==3.0.0
jmespath==0.10.0
joblib==1.0.1
oauthlib==3.1.0
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2020.5
pywin32==227
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
snowflake-connector-python==2.3.10
snowflake-sqlalchemy==1.2.4
SQLAlchemy==1.3.23
urllib3==1.26.4